requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
backcall==0.2.0
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.2
executing==2.0.1
filelock==3.15.4
fsspec==2024.6.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.140.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
grpcio==1.65.4
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.24.5
idna==3.7
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mpmath==1.3.0
networkx==3.3
numpy==2.0.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
openai==1.40.0
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.4
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.1.2
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
rsa==4.9
safetensors==0.4.4
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
sympy==1.13.1
tokenizers==0.19.1
torch==2.4.0
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.0
triton==3.0.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
wcwidth==0.2.13
